03_python_function/weekly_revenue_func/requirements.txt
azure-functions==1.20.0
pyodbc==5.1.0
tzdata

